Move TreeWriter to the correct namespace
